requirements.txt
# AI Core
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.0
openai==1.14.0
tiktoken==0.6.0

# Vector Database
chromadb==0.4.24

# FastAPI
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1

# Data Processing
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.1

# PDF Processing
pypdf==4.1.0

# Utils
python-multipart==0.0.9

# Jupyter (실험용)
notebook==7.0.8
ipykernel==6.29.0

yfinance>=0.2.0
psycopg2-binary>=2.9

